requirements.txt
# TikTok CDN Demo Dependencies
# Compatible with Render.com Python environment

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
aiohttp==3.9.0
jinja2==3.1.2
python-dotenv==1.0.0
